+++ phase1/src/main/java/atm/CreditCard.java
@@ -7,11 +7,11 @@ public class CreditCard extends Account {
         super(owner);
     }
 
-    public int getBalance() {
+    public double getBalance() {
         return this.balance;
     }
 
-    public void setBalance(int amount) {
+    public void setBalance(double amount) {
         this.balance -= amount;
     }
