+++ phase1/src/main/java/atm/LineOfCredit.java
@@ -8,13 +8,13 @@ public class LineOfCredit extends Account {
     }
 
 
-    public int getBalance() {
+    public double getBalance() {
         return this.balance;
     }
 
     // this takes in either a negative balance if money was taken from the account
     // or positive balance if money is added
-    public void setBalance(int amount) { this.balance -= amount; }
+    public void setBalance(double amount) { this.balance -= amount; }
 
     @Override
     public String toString() {
